fully rename NewActivity to ActivityFormg
